DATAJDBC-378 - Polishing
Move non integration tests to separate class.
